pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "biolm-demo"
version = "0.1.0"
description = "BioLM Demo: Biological Language Model demonstrations and experiments"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "torch==2.7.1",
  "transformers==4.55.0",
  "pandas==2.3.1",
  "numpy==2.3.2",
  "scikit-learn==1.7.1",
  "datasets==4.0.0",
  "metaflow==2.17.0",
  "tqdm==4.67.1",
  "pydantic-settings==2.10.1",
  "omegaconf>=2.3.0",
]

[project.optional-dependencies]
mamba = [
  "causal-conv1d==1.5.0.post8",
  "mamba-ssm==2.2.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
causal-conv1d = { git = "https://github.com/Dao-AILab/causal-conv1d", tag = "v1.5.0.post8" }
mamba-ssm = { git = "https://github.com/state-spaces/mamba", tag = "v2.2.4" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true                                # Use this index only for pytorch

[tool.uv]
no-build-isolation-package = ["mamba-ssm", "causal-conv1d"]

[[tool.uv.dependency-metadata]]
name = "mamba-ssm"
version = "2.2.4"
requires-dist = ["torch", "einops"]

[[tool.uv.dependency-metadata]]
name = "causal-conv1d"
version = "1.5.0.post8"
requires-dist = ["torch", "einops"]


[dependency-groups]
dev = [
  "pytest>=8.3.4",
  "ruff>=0.11.13",
  "pytype>=2024.10.11",
  "pre-commit>=4.2.0",
]
slurm = [
  "submitit>=1.5.1",
]
